#12 Separate configuration page cards to files
